Add Node-based tests for editor automation utilities
