+++ .gitignore
@@ -29,6 +29,7 @@ node_modules
 dist
 dist-ssr
 *.local
+.test-build
